requirements.txt
numpy>=1.16.4
matplotlib==2.2.4
Pillow>=6.1.0
scipy>=1.2.2
imageio>=2.5.0
six==1.14.0
tensorflow-datasets==3.2.1
tensorflow-addons==0.11.1
dlib==19.20.0
opencv-python==4.3.0.36
